Added cron

+++ migrations/000001_create_rates_table.up.sql
@@ -4,6 +4,6 @@ CREATE TABLE IF NOT EXISTS exchange_rates (
     cur_abbreviation VARCHAR(10) NOT NULL,
     cur_name VARCHAR(100) NOT NULL,
     cur_scale INT NOT NULL,
-    cur_official_rate INT NOT NULL,
+    cur_official_rate FLOAT NOT NULL,
     PRIMARY KEY (cur_id, date)
 );
